Split weapon loadout into multiple files
